requirements_pretrained.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy
opencv-python
pyyaml

# For pretrained models
timm>=0.9.0

# For distributed training (optional)
submitit

# Additional utilities
pillow
tqdm 